Updated web page, fixed a few bugs
Added a readme with a brief description of what the files are. Added a
few features to the web page. Put the reset time of the warbands in its
own file rather than having it in many places, which also means it can
be updated by the python script now. Added "Access-Control-Allow-Origin"
directive to the main server's config to avoid having to keep a copy of
the reset timer on the mirror.